Remove underscores from package names
See https://blog.golang.org/package-names

Changes are:
gcp_director -> director
service_broker -> servicebroker
stackdriver_nozzle -> strackdrivernozzle
ops_manager -> opsman

Also any files named ops_manager.go were updated to opsman.go for
consistency.

Also, I fixed a bad refactor which was breaking project_test.go
