+++ src/omg-cli/omg/commands/delete_installation.go
@@ -22,7 +22,7 @@ import (
 
 	"omg-cli/config"
 	"omg-cli/omg/setup"
-	"omg-cli/ops_manager"
+	"omg-cli/opsman"
 
 	"github.com/alecthomas/kingpin"
 )
@@ -51,7 +51,7 @@ func (cmd *DeleteInstallationCommand) run(c *kingpin.ParseContext) error {
 		return err
 	}
 
-	omSdk, err := ops_manager.NewSdk(fmt.Sprintf("https://%s", cfg.OpsManagerHostname), cfg.OpsManager, cmd.logger)
+	omSdk, err := opsman.NewSdk(fmt.Sprintf("https://%s", cfg.OpsManagerHostname), cfg.OpsManager, cmd.logger)
 	if err != nil {
 		return err
 	}
