+++ src/omg-cli/omg/commands/director_ssh.go
@@ -26,7 +26,7 @@ import (
 	"path/filepath"
 
 	"omg-cli/config"
-	"omg-cli/ops_manager"
+	"omg-cli/opsman"
 	"omg-cli/version"
 
 	"github.com/alecthomas/kingpin"
@@ -56,7 +56,7 @@ func (cmd *DirectorSSHCommand) run(c *kingpin.ParseContext) error {
 		return err
 	}
 
-	omSdk, err := ops_manager.NewSdk(fmt.Sprintf("https://%s", cfg.OpsManagerHostname), cfg.OpsManager, cmd.logger)
+	omSdk, err := opsman.NewSdk(fmt.Sprintf("https://%s", cfg.OpsManagerHostname), cfg.OpsManager, cmd.logger)
 	if err != nil {
 		return err
 	}
