+++ src/omg-cli/omg/setup/project_test.go
@@ -65,7 +65,7 @@ var _ = Describe("GcpProject", func() {
 			quotaErrors, satisfied, err := project.ValidateQuotas()
 			Expect(err).To(Equal(ErrUnsatisfiedQuota))
 			Expect(quotaErrors).ToNot(BeNil())
-			Expect(quotaErrors).To(ContainElement(quotaError{Quota: quotaRequirement, Region: "global"}))
+			Expect(quotaErrors).To(ContainElement(QuotaError{Quota: quotaRequirement, Region: "global"}))
 			Expect(satisfied).To(BeEmpty())
 		})
 
@@ -103,7 +103,7 @@ var _ = Describe("GcpProject", func() {
 
 			errors, satisfied, err := project.ValidateQuotas()
 			Expect(err).To(Equal(ErrUnsatisfiedQuota))
-			Expect(errors).To(ContainElement(quotaError{Quota: quotaRequirement, Actual: 10.0, Region: "us-east1"}))
+			Expect(errors).To(ContainElement(QuotaError{Quota: quotaRequirement, Actual: 10.0, Region: "us-east1"}))
 			Expect(satisfied).To(BeEmpty())
 		})
 	})
