+++ src/omg-cli/omg/tiles/servicebroker/configuration.go
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package service_broker
+package servicebroker
 
 import (
 	"encoding/json"
 
 	"omg-cli/config"
 	"omg-cli/omg/tiles"
-	"omg-cli/ops_manager"
+	"omg-cli/opsman"
 )
 
 type properties struct {
@@ -32,7 +32,7 @@ type properties struct {
 }
 
 // Configure satisfies TileInstaller interface.
-func (*Tile) Configure(envConfig *config.EnvConfig, cfg *config.Config, om *ops_manager.Sdk) error {
+func (*Tile) Configure(envConfig *config.EnvConfig, cfg *config.Config, om *opsman.Sdk) error {
 	if err := om.StageProduct(tile.Product); err != nil {
 		return err
 	}
