+++ src/omg-cli/omg/tiles/stackdrivernozzle/resource.go
@@ -14,20 +14,41 @@
  * limitations under the License.
  */
 
-package gcp_director
+package stackdrivernozzle
 
 import (
+	"log"
+
 	"omg-cli/config"
 )
 
 var tile = config.Tile{
+	Pivnet: config.PivnetMetadata{
+		Name:      "gcp-stackdriver-nozzle",
+		ReleaseID: 53596,
+		FileID:    89124,
+		Sha256:    "80e137622ca76868693b406114a2c7c1fdf6ce5db91c77a8d848d558d288fe5c",
+	},
 	Product: config.OpsManagerMetadata{
-		Name: "BOSH Director",
+		Name:         "stackdriver-nozzle",
+		Version:      "2.0.1",
+		DependsOnPAS: true,
+	},
+	Stemcell: &config.StemcellMetadata{
+		PivnetMetadata: config.PivnetMetadata{
+			Name:      "stemcells",
+			ReleaseID: 214323,
+			FileID:    247292,
+			Sha256:    "8c6caeae37711aaf12b4fefba06c348cde5631e872e8892553ddb26514a3953a",
+		},
+		StemcellName: "light-bosh-stemcell-3468.78-google-kvm-ubuntu-trusty-go_agent",
 	},
 }
 
-// Tile is the tile for the BOSH Director.
-type Tile struct{}
+// Tile is the tile for the Stackdriver Nozzle.
+type Tile struct {
+	Logger *log.Logger
+}
 
 // Definition satisfies TileInstaller interface.
 func (*Tile) Definition(*config.EnvConfig) config.Tile {
@@ -36,5 +57,5 @@ func (*Tile) Definition(*config.EnvConfig) config.Tile {
 
 // BuiltIn satisfies TileInstaller interface.
 func (*Tile) BuiltIn() bool {
-	return true
+	return false
 }
